dictionary implemented. correction issue with strings and pointers probably

+++ Consegna_2/src/edit_distance_lib.c
@@ -17,8 +17,8 @@ int min(int a, int b, int c) {
 //parte dall'inizio della stringa, per comodità di lettura
 int edit_distance(char *string1, char *string2){
     int no_op;
-    if (strlen(string1) == 0) return strlen(string2);
-    if (strlen(string2) == 0) return strlen(string1);
+    if (strlen(string1) == 0) return (int)strlen(string2);
+    if (strlen(string2) == 0) return (int)strlen(string1);
 
     if (string1[0] == string2[0]){
         no_op = edit_distance(string1+1, string2+1);
